Add function to create/destroy video surfaces without full VDPAU device support
this functionality is used to create video surfaces without a VDPAU context. This functionality is used in kodi for displaying video frames decoded outside of VDPAU (e.g. with FFMPEG software decoder), but the frames are displayed via the allwinner display hardware (i.e. with help of a hardware layer)